move BehaviorAssignment definition into src/polymer
